requirements.txt
streamlit
tensorflow
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
